pyproject.toml
[tool.poetry]
name = "fastapi-template-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10.4"
fastapi = "0.103.2"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.22.0"
SQLAlchemy = "^2.0.21"
pydantic-settings = "^2.0.3"
dependency-injector = "^4.41.0"
asyncpg = "^0.28.0"
starlette-context = "^0.3.6"
cryptography = "41.0.5"
bcrypt = "4.1.3"

[tool.poetry.dev-dependencies]
alembic = "^1.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
